Move EachOwed component to its own folder
