+++ src/EachOwed/EachOwed.css
@@ -1,4 +1,4 @@
-@import url('./global.css');
+@import url('../global.css');
 
 
 .each-owed {
